Make all asViewStore method to AbstractViewStore
Moves all of the `asViewStore` methods out of `AbstractStore` and into an extension on `AbstractViewStore`.
Having these functions operable on `AbstractStore` only unnecessarily limits their scope.  A view store only
deals in State and Action - in both the _from_ store and the _to_ store.  Moving these functions to a context
where that's all that exists frees programmers to map one _view_ store to another.